+++ packages/frontend/src/components/dialogs/CreateChat/index.tsx
@@ -266,6 +266,16 @@ function CreateChatMain(props: CreateChatMainProps) {
 
   const fixedSizeListOuterRef = useRef<HTMLElement>(null)
 
+  const onKeyDown = (ev: React.KeyboardEvent) => {
+    if (ev.code === 'ArrowDown') {
+      ;(
+        fixedSizeListOuterRef.current?.querySelector('button') as HTMLElement
+      )?.focus()
+      // prevent scrolling down the list
+      ev.preventDefault()
+    }
+  }
+
   return (
     <>
       <DialogHeader>
@@ -277,6 +287,7 @@ function CreateChatMain(props: CreateChatMainProps) {
           placeholder={
             isChatmail ? tx('search') : tx('contacts_enter_name_or_email')
           }
+          onKeyDown={onKeyDown}
           autoFocus
           spellCheck={false}
         />